pyproject.toml
[project]
name = "layoutvlm-lab"
version = "0.1.0"
description = "LayoutVLM-Lab: High-concurrency document parsing with Layout + VLM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml==6.0.2",
    "python-dotenv==1.2.1",
    "tqdm==4.67.1",
    "pillow==12.1.0",
    "openai==2.15.0",
    "bs4==0.0.2",
]

[project.optional-dependencies]
paddle = [
  "paddlepaddle-gpu==3.2.0",
  "paddleocr[doc-parser]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"